Solution to Check Greater Than

+++ Check Greater Than@@67882a731b4392d8fc1577d6/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 int main(int argc, char** argv) {
-   int a = argv[1];
-   int b = argv[2];
+   long a = argv[1];
+   long b = argv[2];
     a > b ? printf("True") : printf("False");
     return 0;
 }
